# specs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/audit-logs": {
      "get": {
        "operationId": "AuditLogsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AuditLogs"
        ]
      },
      "post": {
        "operationId": "AuditLogsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAuditLogDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "AuditLogs"
        ]
      }
    },
    "/audit-logs/{id}": {
      "get": {
        "operationId": "AuditLogsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AuditLogs"
        ]
      },
      "patch": {
        "operationId": "AuditLogsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAuditLogDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AuditLogs"
        ]
      },
      "delete": {
        "operationId": "AuditLogsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "AuditLogs"
        ]
      }
    },
...
    "/players/{id}": {
      "get": {
        "operationId": "PlayersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Player details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get a player by ID",
        "tags": [
          "players"
        ]
      },
      "patch": {
        "operationId": "PlayersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlayerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Player updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update a player",
        "tags": [
          "players"
        ]
      },
      "delete": {
        "operationId": "PlayersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Player deleted"
          }
        },
        "summary": "Delete a player",
        "tags": [
          "players"
        ]
      }
    },
...
    "/sensitive-player-data": {
      "get": {
        "operationId": "SensitivePlayerDataController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of sensitive player data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SensitivePlayerDataListItemDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all sensitive player data",
        "tags": [
          "sensitive-player-data"
        ]
      },
      "post": {
        "operationId": "SensitivePlayerDataController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSensitivePlayerDataDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sensitive player data created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensitivePlayerDataResponseDto"
                }
              }
            }
          }
        },
        "summary": "Create new sensitive player data",
        "tags": [
          "sensitive-player-data"
        ]
      }
    },
    "/sensitive-player-data/{id}": {
      "get": {
        "operationId": "SensitivePlayerDataController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sensitive player data details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensitivePlayerDataResponseDto"
                }
              }
            }
          }
        },
        "summary": "Get sensitive player data by ID",
        "tags": [
          "sensitive-player-data"
        ]
      },
      "patch": {
        "operationId": "SensitivePlayerDataController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSensitivePlayerDataDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sensitive player data updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensitivePlayerDataResponseDto"
                }
              }
            }
          }
        },
        "summary": "Update sensitive player data",
        "tags": [
          "sensitive-player-data"
        ]
      },
      "delete": {
        "operationId": "SensitivePlayerDataController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Sensitive player data deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SensitivePlayerDataResponseDto"
                }
              }
            }
          }
        },
        "summary": "Delete sensitive player data",
        "tags": [
          "sensitive-player-data"
        ]
      }
    },
...
    "/staff": {
      "get": {
        "operationId": "StaffController_findAll",
        "parameters": [
          {
            "name": "clubId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Staff"
        ]
      },
      "post": {
        "operationId": "StaffController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStaffDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Staff"
        ]
      }
    },
    "/staff/{id}": {
      "get": {
        "operationId": "StaffController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Staff"
        ]
      },
      "patch": {
        "operationId": "StaffController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStaffDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Staff"
        ]
      },
      "delete": {
        "operationId": "StaffController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Staff"
        ]
      }
    },
...
  },
  "info": {
    "title": "PlayerTracker API",
    "description": "API pour la plateforme de suivi physique et mental des athlètes",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateAuditLogDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "description": "ID de l'utilisateur",
            "example": 1
          },
          "action": {
            "type": "string",
            "description": "Action effectuée",
            "example": "CREATE_PLAYER",
            "maxLength": 255
          },
          "targetType": {
            "type": "string",
            "description": "Type de la cible",
            "example": "Player",
            "maxLength": 100
          },
          "targetId": {
            "type": "number",
            "description": "ID de la cible",
            "example": 1
          }
        },
        "required": [
          "action",
          "targetType"
        ]
      },
      "UpdateAuditLogDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "description": "ID de l'utilisateur",
            "example": 1
          },
          "action": {
            "type": "string",
            "description": "Action effectuée",
            "example": "CREATE_PLAYER",
            "maxLength": 255
          },
          "targetType": {
            "type": "string",
            "description": "Type de la cible",
            "example": "Player",
            "maxLength": 100
          },
          "targetId": {
            "type": "number",
            "description": "ID de la cible",
            "example": 1
          }
        }
      },
...
      "ClubNestedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Club ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Club name",
            "example": "FC Barcelona"
          },
          "address": {
            "type": "object",
            "description": "Club address",
            "example": "123 Main St",
            "nullable": true
          },
          "phone": {
            "type": "object",
            "description": "Phone number",
            "example": "+34 123 456 789",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp",
            "example": "2024-01-01T00:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp",
            "example": "2024-01-15T12:00:00Z"
          }
        },
        "required": [
          "id",
          "name",
          "address",
          "phone",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "PositionNestedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Position ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "Position name",
            "example": "Forward"
          },
          "sportId": {
            "type": "number",
            "description": "Sport ID",
            "example": 1
          }
        },
        "required": [
          "id",
          "name",
          "sportId"
        ]
      },
...
      "UserNestedDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "User ID",
            "example": 1
          },
          "email": {
            "type": "string",
            "description": "Email address",
            "example": "john.doe@example.com"
          },
          "avatarUrl": {
            "type": "object",
            "description": "Avatar URL",
            "example": "https://example.com/avatar.jpg",
            "nullable": true
          },
          "isEmailVerified": {
            "type": "boolean",
            "description": "Is email verified",
            "example": true
          },
          "lastLoginAt": {
            "type": "object",
            "description": "Last login timestamp",
            "example": "2024-01-15T10:30:00Z",
            "nullable": true
          },
          "role": {
            "type": "string",
            "description": "User role",
            "example": "PLAYER",
            "enum": [
              "PLAYER",
              "STAFF",
              "ADMIN"
            ]
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp",
            "example": "2024-01-01T00:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp",
            "example": "2024-01-15T12:00:00Z"
          }
        },
        "required": [
          "id",
          "email",
          "avatarUrl",
          "isEmailVerified",
          "lastLoginAt",
          "role",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "PlayerResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "Player ID",
            "example": 1
          },
          "userId": {
            "type": "number",
            "description": "User ID",
            "example": 1,
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "description": "First name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "Last name",
            "example": "Doe"
          },
          "nickName": {
            "type": "string",
            "description": "Nickname",
            "example": "Johnny",
            "nullable": true
          },
          "photoUrl": {
            "type": "string",
            "description": "Photo URL",
            "example": "https://example.com/photo.jpg",
            "nullable": true
          },
          "clubId": {
            "type": "number",
            "description": "Club ID",
            "example": 1
          },
          "teamId": {
            "type": "number",
            "description": "Team ID",
            "example": 1,
            "nullable": true
          },
          "positionId": {
            "type": "number",
            "description": "Position ID",
            "example": 1,
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "description": "Is active",
            "example": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp",
            "example": "2024-01-01T00:00:00Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp",
            "example": "2024-01-15T12:00:00Z"
          },
          "user": {
            "description": "User details",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/UserNestedDto"
              }
            ]
          },
          "club": {
            "description": "Club details",
            "allOf": [
              {
                "$ref": "#/components/schemas/ClubNestedDto"
              }
            ]
          },
          "team": {
            "description": "Team details",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/TeamNestedDto"
              }
            ]
          },
          "position": {
            "description": "Position details",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/PositionNestedDto"
              }
            ]
          },
          "sensitiveData": {
            "description": "Sensitive player data (PII/health)",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/SensitivePlayerDataNestedDto"
              }
            ]
          }
        },
        "required": [
          "id",
          "userId",
          "firstName",
          "lastName",
          "nickName",
          "photoUrl",
          "clubId",
          "teamId",
          "positionId",
          "isActive",
          "createdAt",
          "updatedAt",
          "user",
          "club",
          "team",
          "position",
          "sensitiveData"
        ]
      },
...
      "UpdatePlayerDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "description": "ID de l'utilisateur lié",
            "example": 1
          },
          "firstName": {
            "type": "string",
            "description": "Prénom du joueur",
            "example": "John",
            "minLength": 1,
            "maxLength": 100
          },
          "lastName": {
            "type": "string",
            "description": "Nom du joueur",
            "example": "Doe",
            "minLength": 1,
            "maxLength": 100
          },
          "nickName": {
            "type": "string",
            "description": "Surnom du joueur",
            "example": "Johnny",
            "minLength": 1,
            "maxLength": 100
          },
          "photoUrl": {
            "type": "string",
            "description": "URL de la photo du joueur",
            "example": "http://example.com/photo.jpg",
            "maxLength": 255
          },
          "clubId": {
            "type": "number",
            "description": "ID du club",
            "example": 1
          },
          "teamId": {
            "type": "number",
            "description": "ID de l'équipe",
            "example": 1
          },
          "positionId": {
            "type": "number",
            "description": "ID de la position",
            "example": 1
          },
          "isActive": {
            "type": "boolean",
            "description": "Joueur actif",
            "example": true
          }
        }
      },
...
      "SensitivePlayerDataListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID of the sensitive player data",
            "example": 1
          },
          "playerId": {
            "type": "number",
            "description": "ID of the player",
            "example": 1
          },
          "weight": {
            "type": "object",
            "description": "Weight in kg",
            "example": 75.5
          },
          "height": {
            "type": "object",
            "description": "Height in cm",
            "example": 180
          },
          "birthDate": {
            "type": "object",
            "description": "Birth date",
            "example": "1995-05-15T00:00:00.000Z"
          },
          "nationality": {
            "type": "object",
            "description": "Nationality (ISO-2 code)",
            "example": "FR"
          },
          "gender": {
            "type": "string",
            "description": "Gender",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ],
            "example": "MALE"
          }
        },
        "required": [
          "id",
          "playerId"
        ]
      },
      "SensitivePlayerDataResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "ID of the sensitive player data",
            "example": 1
          },
          "playerId": {
            "type": "number",
            "description": "ID of the player",
            "example": 1
          },
          "weight": {
            "type": "object",
            "description": "Weight in kg",
            "example": 75.5
          },
          "height": {
            "type": "object",
            "description": "Height in cm",
            "example": 180
          },
          "birthDate": {
            "type": "object",
            "description": "Birth date",
            "example": "1995-05-15T00:00:00.000Z"
          },
          "nationality": {
            "type": "object",
            "description": "Nationality (ISO-2 code)",
            "example": "FR"
          },
          "gender": {
            "type": "string",
            "description": "Gender",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ],
            "example": "MALE"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation date",
            "example": "2024-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update date",
            "example": "2024-01-01T00:00:00.000Z"
          }
        },
        "required": [
          "id",
          "playerId",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateSensitivePlayerDataDto": {
        "type": "object",
        "properties": {
          "playerId": {
            "type": "number",
            "description": "ID du joueur",
            "example": 1
          },
          "weight": {
            "type": "number",
            "description": "Poids (kg)",
            "example": 75.5
          },
          "height": {
            "type": "number",
            "description": "Taille (cm)",
            "example": 180
          },
          "birthDate": {
            "format": "date-time",
            "type": "string",
            "description": "Date de naissance",
            "example": "1995-01-01T00:00:00.000Z"
          },
          "medicalNotes": {
            "type": "string",
            "description": "Notes médicales",
            "maxLength": 2000
          },
          "nationality": {
            "type": "string",
            "description": "Nationalité (code pays ISO-2)",
            "example": "FR"
          },
          "gender": {
            "type": "string",
            "description": "Genre",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ],
            "example": "MALE"
          }
        },
        "required": [
          "playerId",
          "gender"
        ]
      },
      "UpdateSensitivePlayerDataDto": {
        "type": "object",
        "properties": {
          "playerId": {
            "type": "number",
            "description": "ID du joueur",
            "example": 1
          },
          "weight": {
            "type": "number",
            "description": "Poids (kg)",
            "example": 75.5
          },
          "height": {
            "type": "number",
            "description": "Taille (cm)",
            "example": 180
          },
          "birthDate": {
            "format": "date-time",
            "type": "string",
            "description": "Date de naissance",
            "example": "1995-01-01T00:00:00.000Z"
          },
          "medicalNotes": {
            "type": "string",
            "description": "Notes médicales",
            "maxLength": 2000
          },
          "nationality": {
            "type": "string",
            "description": "Nationalité (code pays ISO-2)",
            "example": "FR"
          },
          "gender": {
            "type": "string",
            "description": "Genre",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ],
            "example": "MALE"
          }
        }
      },
...
      "CreateStaffDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "description": "ID de l'utilisateur",
            "example": 1
          },
          "clubId": {
            "type": "number",
            "description": "ID du club",
            "example": 1
          },
          "firstName": {
            "type": "string",
            "description": "Prénom",
            "example": "Jean",
            "maxLength": 100
          },
          "lastName": {
            "type": "string",
            "description": "Nom",
            "example": "Dupont",
            "maxLength": 100
          },
          "specialty": {
            "type": "string",
            "description": "Spécialité",
            "example": "Préparateur physique",
            "maxLength": 255
          },
          "phone": {
            "type": "string",
            "description": "Téléphone",
            "example": "+33612345678",
            "maxLength": 50
          },
          "certifications": {
            "type": "string",
            "description": "Certifications",
            "example": "BEES 2, Diplôme de kinésithérapie",
            "maxLength": 1000
          }
        },
        "required": [
          "userId",
          "clubId"
        ]
      },
      "UpdateStaffDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "description": "ID de l'utilisateur",
            "example": 1
          },
          "clubId": {
            "type": "number",
            "description": "ID du club",
            "example": 1
          },
          "firstName": {
            "type": "string",
            "description": "Prénom",
            "example": "Jean",
            "maxLength": 100
          },
          "lastName": {
            "type": "string",
            "description": "Nom",
            "example": "Dupont",
            "maxLength": 100
          },
          "specialty": {
            "type": "string",
            "description": "Spécialité",
            "example": "Préparateur physique",
            "maxLength": 255
          },
          "phone": {
            "type": "string",
            "description": "Téléphone",
            "example": "+33612345678",
            "maxLength": 50
          },
          "certifications": {
            "type": "string",
            "description": "Certifications",
            "example": "BEES 2, Diplôme de kinésithérapie",
            "maxLength": 1000
          }
        }
      },
...
    }
  }
}